% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{download}
\alias{download}
\title{Title Download}
\usage{
download(url)
}
\arguments{
\item{url}{The URL with JSON content to download}
}
\value{
A list including data on all amendments;
the total number of amendments; and the number of amendments per page
}
\description{
Title Download
}
