% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bills_api.R
\name{get_bill_stages_ids}
\alias{get_bill_stages_ids}
\title{Get bill stages IDs for a bill}
\usage{
get_bill_stages_ids(bill_id)
}
\arguments{
\item{bill_id}{The unique ID of a bill}
}
\value{
A dataframe with relevant bill stages and their IDs
}
\description{
Get bill stages IDs for a bill
}
