% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bills_api.R
\name{get_ping_pong_items}
\alias{get_ping_pong_items}
\title{Title Get ping pong items}
\usage{
get_ping_pong_items(bill_id)
}
\arguments{
\item{bill_id}{The unique ID of a bill}
}
\value{
A tibble with detail, but not content, of ping pong items
}
\description{
Title Get ping pong items
}
