% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_sponsor}
\alias{get_sponsor}
\title{Title Get sponsor}
\usage{
get_sponsor(list_element)
}
\arguments{
\item{list_element}{The dataframe that contains sponsor data}
}
\value{
A tibble with key data for the lead sponsor
}
\description{
Title Get sponsor
}
