% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bills_api.R
\name{get_stage_amendments}
\alias{get_stage_amendments}
\title{Get stage amendments}
\usage{
get_stage_amendments(bill_id, bill_stage_id)
}
\arguments{
\item{bill_id}{The unique ID of a bill}

\item{bill_stage_id}{The unique ID of a bill stage}
}
\value{
A dataframe with all amendments for a bill stage
}
\description{
Get stage amendments
}
