% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ministers_stages.R
\name{get_stage_ministers}
\alias{get_stage_ministers}
\title{Title Get stage ministers}
\usage{
get_stage_ministers(bill_id, bill_stage_id)
}
\arguments{
\item{bill_id}{The unique ID of a bill}

\item{bill_stage_id}{The unique ID of a bill stage}
}
\value{
A tibble with one row per ministerial role held during the bill stage
}
\description{
Title Get stage ministers
}
