% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gvt_amd.R
\name{get_total_gvt_amendments}
\alias{get_total_gvt_amendments}
\title{Title Get total government amendments}
\usage{
get_total_gvt_amendments(bill_id)
}
\arguments{
\item{bill_id}{The unique ID of a bill}
}
\value{
A list including a dataframe with all government amendments and one with ping pong items
}
\description{
Title Get total government amendments
}
