% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gvt_amd.R
\name{stage_gvt_amd}
\alias{stage_gvt_amd}
\title{Stage government amendments}
\usage{
stage_gvt_amd(bill_id, bill_stage_id)
}
\arguments{
\item{bill_id}{The unique ID of a bill}

\item{bill_stage_id}{The unique ID of a bill stage}
}
\value{
A dataframe with amendments introduced during a bill stage by ministers
}
\description{
Stage government amendments
}
